+++ app/src/main/java/com/luckyzyx/luckytool/hook/hooker/BatteryInfoNotify.kt
@@ -14,10 +14,14 @@ import java.util.*
 import kotlin.math.abs
 
 object BatteryInfoNotify : YukiBaseHooker() {
+    //battery
     private var plugged: String = ""
     private var temperature: Double = 0.0
     private var voltage: Double = 0.0
     private var electricCurrent: Int = 0
+    private var max_charging_current: Int = 0
+    private var max_charging_voltage: Int = 0
+    //oplus battery
     private var chargerVoltage: Double = 0.0
     private var chargerTechnology: Int = 0
     override fun onHook() {
@@ -68,14 +72,16 @@ object BatteryInfoNotify : YukiBaseHooker() {
             else -> "null"
         }
         temperature = (intent.getIntExtra(BatteryManager.EXTRA_TEMPERATURE, 0) / 10.0)
-        voltage = (intent.getIntExtra(BatteryManager.EXTRA_VOLTAGE, 0) / 1000.0)
+        val originalVoltage = intent.getIntExtra(BatteryManager.EXTRA_VOLTAGE, 0)
+        voltage = if (originalVoltage.toString().length == 1) originalVoltage * 1.0 else originalVoltage / 1000.0
         electricCurrent = batteryManager.getIntProperty(BatteryManager.BATTERY_PROPERTY_CURRENT_NOW)
+        max_charging_current = intent.getIntExtra("max_charging_current", 0)
+        max_charging_voltage = intent.getIntExtra("max_charging_voltage", 0)
     }
 
     private fun initOplusInfo(intent: Intent) {
         chargerVoltage = (intent.getIntExtra("chargervoltage", 0) / 1000.0)
         chargerTechnology = (intent.getIntExtra("chargertechnology",0))
-        //chargefastcharger
     }
 
     private fun createChannel(context: Context) {
@@ -95,35 +101,39 @@ object BatteryInfoNotify : YukiBaseHooker() {
             return
         }
         createChannel(context)
-
         val defaultInfo = if (isZh(context)) {
             "温度:${temperature}℃ 电压:${voltage}v 电流:${electricCurrent}mA"
         } else "${temperature}℃ ${voltage}v ${electricCurrent}mA"
-
         val technology = when(chargerTechnology){
             0 -> "Normal"
             1 -> "Vooc"
             2 -> "SuperVooc 5A"
             20 -> "SuperVooc2 6.5A"
-            30 -> "SuperVooc Athena Foregin Pro"
+            30 -> "SuperVooc Athena Foreign Pro"
             25 -> "Vooc Beta Pro"
             3 -> "PD"
             4 -> "QC"
             else -> "Error"
         }
+        val chargerVoltageFinal = when (chargerTechnology) {
+            0 -> if (max_charging_current == 0 && max_charging_voltage == 0) chargerVoltage else 5.0
+            //vooc,pd,qc
+            1, 3, 4 -> chargerVoltage
+            //svooc
+            2, 20, 25, 30 -> voltage * 2
+            else -> 0.0
+        }
         val chargeInfo = if (isCharge) {
-            val power = Formatter().format("%.2f", (chargerVoltage * abs(electricCurrent)) / 1000.0).toString()
+            val power = Formatter().format("%.2f", (chargerVoltageFinal * abs(electricCurrent)) / 1000.0).toString()
             if (isZh(context)) {
-                "充电中:$plugged 充电电压:${chargerVoltage}v 理论功率:${power}W\n充电技术:${technology}" + if (isUpdateTime) "\n" else ""
+                "充电中:$plugged 充电电压:${chargerVoltageFinal}v 理论功率:${power}W\n充电技术:${technology}" + if (isUpdateTime) "\n" else ""
             } else "$plugged ${chargerVoltage}v ${power}W $technology" + if (isUpdateTime) "\n" else ""
         } else ""
-
         val updateTime = if (isUpdateTime) {
             if (isZh(context)) {
                 "更新时间:${SimpleDateFormat("yyyy/MM/dd HH:mm:ss").format(Date())}"
             } else SimpleDateFormat("MM/dd/yyyy HH:mm:ss").format(Date())
         } else ""
-
         val notify = NotificationCompat.Builder(context, "luckytool_notify").apply {
             setAutoCancel(false)
             setOngoing(true)